When a vehicle has to catch up, don't beeline
Find the first timepoint that's in the future, then edit the timepoints
between the current point and that timepoint so that their 'time'
properties are proportional to the distance that needs to be traveled to
catch up to the prediction. The result is that buses that are behind the
current prediction should no longer beeline to their next position.