Stubbed out code for adjusting route paths that overlap
The logic is simple. Some part of the application (presumably the map)
is responsible for keeping track of all the line segments that are
currently drawn. When a new route is added, check each of the line
segments that make up its path. If a line segment in the path already
exists, shift the endpoints of that segment along a vector normal to the
original segment. If that fails, try -1 * the vector, then 2 * the
vector, and so on until an unoccupied segment is found. Update the
segment "set" and append the points in the adjusted segment to the
output path.