An attempt to improve some aspects of route path offsetting
The first time a path is offset, store a 'scaler' function that takes the
iteration of the path offsetter and calculates the (scalar) magnitude of the
normal. The goal is to consistently offset routes to the same side of the
street, and it works because the order in which coordinates are specified. It
still breaks down on loops, but they are likely to require some special
handling.

This also ought to reduce the overall number of path segment offset trials.